# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的记账系统 API，支持用户注册、登录、消费记录管理和数据导出功能",
        "title": "记账系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/ai-models": {
            "get": {
                "description": "获取系统中所有AI模型配置列表（不包含APIKey）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "获取AI模型列表",
                "responses": {
                    "200": {
                        "description": "获取成功，返回模型列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的AI模型配置，包括名称、API地址和密钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "创建AI模型",
                "parameters": [
                    {
                        "description": "AI模型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateAIModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或模型名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/reorder": {
            "put": {
                "description": "根据传入的模型ID顺序更新排序，用于前端拖拽排序后保存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "排序AI模型",
                "parameters": [
                    {
                        "description": "模型ID顺序",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReorderAIModelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "排序成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/{id}": {
            "get": {
                "description": "根据ID获取AI模型配置详情（不包含APIKey）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "获取单个AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定的AI模型配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "更新AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的模型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateAIModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或模型名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的AI模型配置（软删除）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "删除AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/{id}/test": {
            "post": {
                "description": "向AI模型发送轻量测试请求，检测接口是否可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "检测AI接口可用性",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "检测成功，接口可用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "接口不可用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/ai-models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统可用的AI模型配置列表（不包含APIKey），用于前端选择模型。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "获取AI模型列表",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.AIModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CreateAIModelRequest": {
            "type": "object",
            "required": [
                "api_key",
                "base_url",
                "name"
            ],
            "properties": {
                "api_key": {
                    "type": "string",
                    "minLength": 1,
                    "example": "sk-..."
                },
                "base_url": {
                    "type": "string",
                    "example": "https://api.openai.com/v1"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "OpenAI GPT-4"
                }
            }
        },
...
        "api.ReorderAIModelsRequest": {
            "type": "object",
            "required": [
                "model_ids"
            ],
            "properties": {
                "model_ids": {
                    "description": "按新顺序排列的模型 ID 列表",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateAIModelRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "minLength": 1
                },
                "base_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "models.AIModel": {
            "type": "object",
            "properties": {
                "base_url": {
                    "description": "调用地址",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "sort_order": {
                    "description": "排序序号，越小越靠前",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}